requirements.txt
Django==4.1.7
djangorestframework==3.14.0
networkx==3.0
pandas==1.5.3